OrchidOs Rotation : Configurable 0, 90, 180 and 270 degree rotation [1/2]
Change-Id: Ia1859c51d71ef9d01cec2d13c6468ed89c6ac53e
Contributor: jonasl

- Add system setting

Author: Tim Schumacher <timschumi@gmx.de>
Date:   Wed Nov 28 21:44:18 2018 +0100

    Add back ACCELEROMETER_ROTATION_ANGLES and update references

    This needed to move back into fw/b, because it needs to be
    accessed from inside the RotationPolicy. Previously
    (when this constant and Setting were located in lineage-sdk),
    the settings for the display rotation had no effect, because we
    couldn't query the correct set of settings.

OrchidOs Playstation : Require CONFIG_HID_PLAYSTATION for DualSense key layout

Some olders kernels don't yet support HID_PLAYSTATION. While they are
working on supporting this driver, provide this temporary solution.

Before this patch, if the key layout is loaded, but there is no
HID_PLAYSTATION, then the mappings on the device will be incorrect.

With this patch, this key layout will not be loaded unless there is a
CONFIG_HID_PLAYSTATION enabled on the device.

Instead, when the main layout cannot be loaded, we make another attempt
to load a fallback layout. This way, the mappings on DualSense can be
correct even if the HID_PLAYSTATION module is not loaded.

In this CL, the fallback layout was tested using controllerTester.

Bug: 228005926
Test: tested mappings with a DualSense controller on a device that does
not have CONFIG_HID_PLAYSTATION

Merged-In: I7e46ae1943cd2c63b8c0b03fdf5d71adeea89f6f
Change-Id: I7e46ae1943cd2c63b8c0b03fdf5d71adeea89f6f
Signed-off-by: aswin7469 <aswinas@pixysos.com>